backend/requirements-thirdparty.txt
ayx_python_sdk==2.2.0
cryptography==41.0.4
pyjwt==2.8.0